Hoist get-line-endpoints and clipping
